requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-openlineage==2.2.0
apache-airflow-providers-postgres
apache-airflow-providers-common-sql>=1.23.0
apache-airflow-providers-common-compat>=1.5.0
attrs>=22.2
openlineage-integration-common>=1.31.0
openlineage-python>=1.31.0
psycopg2-binary>=2.9.9
asyncpg>=0.30.0
apache-airflow-providers-apache-spark==5.2.1
pyspark>=3.1.3
grpcio-status>=1.59.0
# apache-airflow-providers-apache-hdfs==4.8.1
# hdfs[avro,dataframe,kerberos]>=2.5.4
# pandas>=2.1.2,<2.2